+++ src/common/containers/Set.ipp
@@ -22,12 +22,12 @@
 // **************************************************************
 //! \param elt is the element of type T to insert.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
 void Set<T,N,Block>::append(T const& elt)
 {
   // Next element
-  m_subindex = MODULO(m_subindex + 1U, M);
+  m_subindex = MODULO(m_subindex + 1ul, M);
   m_index += !m_subindex;
   ++m_last;
 
@@ -35,7 +35,7 @@ void Set<T,N,Block>::append(T const& elt)
   // occupied.
   if (m_index >= IContainer<T,N,Block>::m_allocated_blocks)
     {
-      IContainer<T,N,Block>::reserveBlocks(1U);
+      IContainer<T,N,Block>::reserveBlocks(1ul);
     }
 
   // Insert element and add the 'Occupied' flag
@@ -51,9 +51,9 @@ void Set<T,N,Block>::append(T const& elt)
 //! \param nth the n'th element (index) of the IContainer we want
 //! to remove.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-void Set<T,N,Block>::remove(const uint32_t nth)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+void Set<T,N,Block>::remove(const size_t nth)
 {
   if (outofbound(nth))
     return ;
@@ -61,8 +61,8 @@ void Set<T,N,Block>::remove(const uint32_t nth)
   // Replace the nth 'th element by the last inserted element
   if (nth != m_last)
     {
-      const uint32_t id = nth / M;
-      const uint32_t sid = MODULO(nth, M);
+      const size_t id = nth / M;
+      const size_t sid = MODULO(nth, M);
 
       IContainer<T,N,Block>::m_blocks[id]->m_block[sid] =
         IContainer<T,N,Block>::m_blocks[m_index]->m_block[m_subindex];
@@ -78,8 +78,8 @@ void Set<T,N,Block>::remove(const uint32_t nth)
 // **************************************************************
 //!
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
 void Set<T,N,Block>::removeLast()
 {
   // Empty the last inserted element
@@ -89,7 +89,7 @@ void Set<T,N,Block>::removeLast()
   // IContainer<T,N,Block>::m_blocks[m_index]->addPendingData(m_subindex, m_subindex);
 
   // Restore index
-  if ((0 == m_last) || (INITIAL_INDEX == m_last))
+  if ((0ul == m_last) || (INITIAL_INDEX == m_last))
     {
       m_index = m_subindex = m_last = INITIAL_INDEX;
     }
@@ -109,9 +109,9 @@ void Set<T,N,Block>::removeLast()
 //! \param index1 index of the 1st element to swap.
 //! \param index2 index of the 2nd element to swap.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-bool Set<T,N,Block>::swap(const uint32_t index1, const uint32_t index2)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+bool Set<T,N,Block>::swap(const size_t index1, const size_t index2)
 {
   // Do not swapt itself
   if (index1 == index2)
@@ -121,10 +121,10 @@ bool Set<T,N,Block>::swap(const uint32_t index1, const uint32_t index2)
   if (outofbound(index2))
     return false;
 
-  const uint32_t id1 = index1 / M;
-  const uint32_t sid1 = MODULO(index1, M);
-  const uint32_t id2 = index2 / M;
-  const uint32_t sid2 = MODULO(index2, M);
+  const size_t id1 = index1 / M;
+  const size_t sid1 = MODULO(index1, M);
+  const size_t id2 = index2 / M;
+  const size_t sid2 = MODULO(index2, M);
 
   T elt = IContainer<T,N,Block>::m_blocks[id2]->m_block[sid2];
   IContainer<T,N,Block>::m_blocks[id2]->m_block[sid2] = IContainer<T,N,Block>::m_blocks[id1]->m_block[sid1];
